test: update framework for react app to delay startup until configured.  Also fix url bug
